# Fix the handling of non-flat intermediate vectors in AverageAggregate