Summary:
AverageAggregate::addIntermediateResults and addSingleGroupIntermediateResults
assume that the fields in the intermediate result RowVector are all flat. This is not true
when the companion functions, e.g., avg_merge and avg_merge_extract_double, are
used by themselves. This diff fixes this issue by letting AverageAggregate::addIntermediateResults
and addSingleGroupIntermediateResults decode the fields of the input vector.

The cost of decoding field vectors is negligible. Below is the benchmark result before and after this fix.
Before:
```
============================================================================
[...]gates/benchmarks/AverageBenchmark.cpp     relative  time/iter   iters/s
============================================================================
flat                                                      194.17ms      5.15
constant                                        152.85%   127.03ms      7.87
dictionary                                      78.342%   247.85ms      4.03
```

After:
```
============================================================================
[...]gates/benchmarks/AverageBenchmark.cpp     relative  time/iter   iters/s
============================================================================
flat                                                      192.22ms      5.20
constant                                        149.39%   128.67ms      7.77
dictionary                                      77.426%   248.26ms      4.03
```

Differential Revision: D47422668